requirements.txt
enhanced-dpy[voice]
google-cloud-firestore
python_dotenv
discord-py-slash-command
henostools
hcolours
jishaku